+++ cg_hermes/config/mutant.py
@@ -11,18 +11,18 @@ MUTANT_COMMON_TAGS = {
         "used_by": ["deliver"],
     },
     frozenset({"variants", "variant-calling"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["tsv", "vcf-report"],
         "used_by": ["deliver"],
     },
     frozenset({"SARS-CoV-2-qc", "result_aggregation"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["qc-report", "csv", "visualization"],
         "used_by": ["deliver"],
     },
     frozenset({"SARS-CoV-2-type", "typing"}): {
         "is_mandatory": False,
-        "tags": ["typing-report", "pangolin", "visualization"],
+        "tags": ["typing-report", "visualization"],
         "used_by": ["deliver"],
     },
     frozenset({"SARS-CoV-2-json", "result_aggregation"}): {
@@ -31,22 +31,22 @@ MUTANT_COMMON_TAGS = {
         "used_by": ["deliver"],
     },
     frozenset({"SARS-CoV-2-sum", "report"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["summary", "csv", "visualization"],
         "used_by": ["deliver"],
     },
     frozenset({"SARS-CoV-2-var", "report"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["variants", "csv", "visualization"],
         "used_by": ["deliver"],
     },
     frozenset({"SARS-CoV-2-info", "report"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["komplettering", "fohm", "visualization"],
         "used_by": ["deliver"],
     },
     frozenset({"runtime-settings", "runinfo"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["mutant-config"],
         "used_by": ["storage", "audit"],
     },
@@ -61,12 +61,12 @@ MUTANT_COMMON_TAGS = {
         "used_by": ["storage"],
     },
     frozenset({"consensus"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["fastq", "consensus"],
         "used_by": ["deliver", "storage"],
     },
     frozenset({"runinfo", "logfile"}): {
-        "is_mandatory": True,
+        "is_mandatory": False,
         "tags": ["mutant-log"],
         "used_by": ["audit"],
     },
@@ -74,5 +74,20 @@ MUTANT_COMMON_TAGS = {
         "is_mandatory": False,
         "tags": ["bam"],
         "used_by": ["storage"],
+    },
+    frozenset({"multiqc-json"}): {
+        "is_mandatory": False,
+        "tags": ["multiqc-json"],
+        "used_by": ["vogue"],
+    },
+    frozenset({"pangolin-typing"}): {
+        "is_mandatory": False,
+        "tags": ["pangolin", "typing-report", "csv"],
+        "used_by": ["deliver"],
+    },
+    frozenset({"pangolin-summary"}): {
+        "is_mandatory": False,
+        "tags": ["pangolin", "typing-summary", "csv"],
+        "used_by": ["deliver"],
     },
 }