+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+## [0.6.0]
+### Changed
+- Mutant output files made optional. 
+### Added  
+- New optional/placeholder tags for pangolin delivery reports.
+- New optional/placeholder tags for multiqc report
+
 ## [0.5.0]
 ### Changed
 - Pangolin output from mutant is optional